Support Unix socket listener
This removes the "port" directive in favor of "bind", and for
consistency also exchanges the "PORT" variable with "BIND".
This allows users to override the default TCP listener with not only
a different port, but also optionally with a Unix socket, by setting
a "BIND" environment variable, without having to modify puma.rb.

The default listener is additionally changed to localhost to improve
security in installations without customization.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>